requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-environ==0.11.2
django-filter==25.1
django-redis==5.4.0
django-simple-history==3.8.0
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
h11==0.16.0
hiredis==3.1.0
httptools==0.6.4
humanize==4.12.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.2
Markdown==3.8
MarkupSafe==3.0.2
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg==3.2.6
psycopg-binary==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydot==3.0.4
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
text-unidecode==1.3
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.35.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
zope.interface==7.2
python-json-logger==2.0.7
django-prometheus==2.2.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-django==0.43b0
opentelemetry-exporter-otlp==1.22.0
python-magic==0.4.27
django-ratelimit==4.1.0
django-ckeditor-5==0.2.18
django-jalali-date==2.0.0
pillow-avif-plugin==1.5.2
Faker==19.3.1
